// types/project.ts
export type Project = {
  _id: string;
  _updatedAt: string;
  title?: string;
  shortTitle?: string;
  titleSs01Map?: string;
  slug?: any;
  description?: any;
  thumbnail?: any;
  gallery?: any[];
  archive?: boolean;
  tags?: any;
  type?: any;
  project?: any;
  role?: any;
  workingAs?: any;
};